GO-6147: UpdateStatus queue

+++ core/files/reconciler/reconciler.go
@@ -73,7 +73,7 @@ func (r *reconciler) Init(a *app.App) error {
 	r.fileSync = app.MustComponent[filesync.FileSync](a)
 	r.fileStorage = app.MustComponent[filestorage.FileStorage](a)
 
-	r.fileSync.OnUploaded(r.markAsReconciled)
+	r.fileSync.OnStatusUpdated(r.markAsReconciled)
 
 	provider := app.MustComponent[anystoreprovider.Provider](a)
 	db := provider.GetCommonDb()
@@ -179,11 +179,14 @@ func (r *reconciler) rebindHandler(ctx context.Context, item *queueItem) (persis
 	return persistentqueue.ActionDone, nil
 }
 
-func (r *reconciler) markAsReconciled(fileObjectId string, fileId domain.FullFileId) error {
+func (r *reconciler) markAsReconciled(fileObjectId string, fileId domain.FullFileId, status filesyncstatus.Status) error {
 	if !r.isRunning() {
 		return nil
 	}
-	return r.deletedFiles.Delete(context.Background(), fileId.FileId.String())
+	if status == filesyncstatus.Synced {
+		return r.deletedFiles.Delete(context.Background(), fileId.FileId.String())
+	}
+	return nil
 }
 
 func (r *reconciler) reconcileRemoteStorage(ctx context.Context) error {
